Major changes: customize field name, return array by default
Field name is now specified when calling the macro, or it defaults to
'tags':

   taggable :keywords

We're no longer using a field name different from the accessors --
when modifying the tags field a string argument will automatically be
converted to an array. We always store an array and always get one back
from the reader. It is the view's (or helper's) job to format tags for
display.

   article.keywords = ['opinion', 'politics']
   article.keywords.should == %w[ opinion politics ]
   article.keywords = 'sports, lacrosse'
   article.keywords.should == %w[ sports lacrosse ]